requirements.txt
# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
albumentations>=1.3.0
pandas>=2.0.0

# Evaluation & visualization
matplotlib>=3.7.0
scikit-learn>=1.3.0
tqdm>=4.65.0

# Rendering (Blender scripts)
# Install separately: pip install blenderproc
# Requires Blender 3.x+

# Object Detection - YOLOv5/YOLOv8
# pip install ultralytics

# Object Detection - MMDetection
# pip install -U openmim
# mim install mmengine mmcv mmdet

# Deep Learning (required for training & evaluation)
# torch>=2.0.0
# torchvision>=0.15.0
